Added current (at present) build of modded Unifont
Also includes PNG version. This commit MAY become obsolete between now and mid-September. When Unifont 15.1.x releases in September (Unicode 15.1 happens on September 12th, Unifont 15.0.01 took a day to release after Unicode 15.0, but since Unicode 15.1 is light on added characters, and the lions' share of the added characters aren't easy to draw in a bitmap font of Unifont's size, and the new characters that *aren't* those are Ideographic Description Characters, it is very, very, likely that Unifont 15.1 will release on the same day as Unicode 15.1. 
My mergers typically take 1 hour to complete. Also, FontForge is a VERY unstable program that makes more comprehensive edits a pain because you have to keep mitigating crashes. Once I finish this project, a form of retirement will happen, and after I make a proper download page and such for it, this 6-year project will finally be done, at which point I will work on other things.