Added vertical metrics to fix Pango
According to Inkscape, some versions of Pango will COMPLETELY error out when dealing with vertical CJK if given a font with no vertical metrics. Once again, this font is a prototype. A version bump COULD happen if upstream Unifont goes to 15.0.07 or possibly more, depending on how the next month and a half goes. OR, no such version could ever happen, and this prototype would just need to be given a different name to keep applications happy. I have a few ideas. I also plan on making a proper page and repo for it when the time rolls around and I am able to spare some time for it. Note that this is a fork.