Update MC Resource Pack build to support 1.20.4
Mojang, in their infinite wisdom, has made it so that Unifont 15.1.04 will be used in 1.20.5, going as far as to do a .hex setup, but not addressing the lack of Plane 1 (which includes emoji). As such, I must keep updating the resource pack to support newer MC versions. They are also planning on doing a Unifont-JP toggle. Since UnifontEX is based on Unifont 15.0.06-JP and 11.0.01 Upper (best one can do to merge Plane0+Plane1), I have to specify it for both. Chinese Github users have not complained about the glyph shapes. This is technically upcoming. I duplicated the JSON. I also had to set a version for "Force Unicode Font" users (Uniform), so there's 3 identical JSONs that all point to the same TTF, and I had to set the pack format to 22. The JP presence should satisfy 1.20.5 when it releases. I'm probably going to only update the pack version on non-snapshot releases.